backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
environs
sqlparse
psycopg2-binary
python-dotenv